Update all feed services to ts
