--- packages/client/src/services/feeds/toshl.js
@@ -1,16 +1,0 @@
-import api from '../api'
-
-const getOauthUrl = () =>
-  api.get({ endpoint: 'feed-auth', params: { feed: 'toshl' } })
-const authenticate = (payload) =>
-  api.post({ endpoint: 'feed-auth', payload, params: { feed: 'toshl' } })
-const deauthorize = () =>
-  api.delete({ endpoint: 'feed-auth', params: { feed: 'toshl' } })
-
-const toshlService = {
-  getOauthUrl,
-  authenticate,
-  deauthorize,
-}
-
-export default toshlService
