+++ packages/client/src/services/feeds/types.ts
@@ -1,0 +1,3 @@
+export interface AuthenticatePayload {
+  code: string
+}
