--- packages/client/src/services/feeds/uber.js
@@ -1,16 +1,0 @@
-import api from '../api'
-
-const getOauthUrl = () =>
-  api.get({ endpoint: 'feed-auth', params: { feed: 'uber' } })
-const authenticate = (payload) =>
-  api.post({ endpoint: 'feed-auth', payload, params: { feed: 'uber' } })
-const deauthorize = () =>
-  api.delete({ endpoint: 'feed-auth', params: { feed: 'uber' } })
-
-const uberService = {
-  getOauthUrl,
-  authenticate,
-  deauthorize,
-}
-
-export default uberService
